+++ AudioVideoLib/Tags/IAudioTagWriter.cs
@@ -1,8 +1,0 @@
-namespace AudioVideoLib.Tags;
-
-/// <summary>
-/// Interface for audio tag writers.
-/// </summary>
-public interface IAudioTagWriter
-{
-}